# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Wewelcom - Restaurantes y Autenticación",
        "description": "API para la gestión de restaurantes y autenticación de usuarios en la plataforma Wewelcom.",
        "contact": {
            "email": "danielpalenzuelamora90@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor de Desarrollo Local"
        },
        {
            "url": "https://tudominio.com/api/v1",
            "description": "Servidor de Producción"
        }
    ],
    "paths": {
...
        "/restaurantes": {
            "get": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Obtiene todos los restaurantes",
                "operationId": "0d8a8e04bc46b5f8a156641ef802c8f4",
                "responses": {
                    "200": {
                        "description": "Listado de restaurantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "restaurantes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Restaurante"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Crear un nuevo restaurante",
                "operationId": "b20fbf63286a1a852e72d2bc3d347900",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Restaurante"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Restaurante creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Restaurante creado con éxito"
                                        },
                                        "restaurante": {
                                            "$ref": "#/components/schemas/Restaurante"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la creación del restaurante"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/restaurantes/{nombre}": {
            "get": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Buscar restaurante por nombre",
                "operationId": "0421e1ff41742f0c5b8b06c1629d4655",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "description": "Nombre (parcial o completo) del restaurante a buscar",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Nuevo Restaurante"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de restaurantes encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Restaurante"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró ningún restaurante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontró el restaurante con el nombre Nuevo Restaurante"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/restaurantes/{id}": {
            "put": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Actualizar restaurante",
                "operationId": "40da1ff3ea6ad2c66ccb9d371fe69db7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del restaurante a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Restaurante"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restaurante actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Restaurante actualizado con éxito"
                                        },
                                        "restaurante": {
                                            "$ref": "#/components/schemas/Restaurante"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Restaurante no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontró el restaurante con el id 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la actualización del restaurante"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Eliminar restaurante",
                "operationId": "05047e94e410d49fba72478c3ee811c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del restaurante a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurante eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Restaurante con ID 1 eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Restaurante no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se encontró el restaurante con el id 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Restaurante": {
                "title": "Restaurante",
                "description": "Modelo de restaurante",
                "properties": {
                    "id_restaurante": {
                        "description": "Class Restaurante",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nombre_restaurante": {
                        "type": "string",
                        "example": "La Parrilla Feliz"
                    },
                    "direccion_restaurante": {
                        "type": "string",
                        "example": "Calle Falsa 123, Ciudad"
                    },
                    "telefono_restaurante": {
                        "type": "string",
                        "example": "+34 123 456 789"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticación. Introduce 'Bearer ' seguido de tu token.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKey": {
                "type": "apiKey",
                "description": "API Key para acceder a recursos protegidos.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}